Move calculateNextDecision inside DeedManager
